Add average background color estimation [WIP]
The intent is to mask the label out of the full screenshot of
its rendered frame and compute the average of the remaining
pixels. This is implemented but requires additional testing, in
particular around partial-alpha pixels around the fringe of the
label text. In many cases, this gives a comparable color as the
existing estimate of averaging the four corner points.

The inverse is also attempted, masking out the
background of the screen capture and leaving the text pixels.
This is not working yet, but left in to spur progress. It is not
currently used in computations, only in report output.

Other approaches for both of these should be considered.